Allow HTTP auth protocols to be configurable by way of a "http-auth-types" parameter.
Neon 0.26 added support for selecting the auth protocols to use.  This
change takes advantage of that support to allow the user to specify
the auth protocols they want to use if they have a new enough Neon.

The default protocols if no config is specified are "basic" and
"digest" for HTTP, or "basic", "digest" and "negotiate" for HTTPS.
"negotiate" will typically use GSSAPI/SSPI on Windows, enabling users
-- on a secure network! -- to activate Windows auth over HTTP (the
lack-thereof being a common gotcha reported by users of Neon 0.26.1).

This change does not provide an implementation for ra_serf.

* subversion/include/svn_config.h
  Add SVN_CONFIG_OPTION_HTTP_AUTH_TYPES symbol to the list of
  configuration parameters available for servers.

* subversion/libsvn_subr/config_file.c
  (svn_config_ensure): Write comments on the new "http-auth-types"
   parameter to the config file (if it doesn't already exist).

* subversion/libsvn_ra_dav/session.c
  (get_server_settings): Add neon_auth_types parameter and fill it in
   if we have a new enough Neon, and it has been set in the config file.

  (svn_ra_dav__open): If we have a new enough Neon, respect the HTTP
   auth types configuration, and use its list to specify the auth
   protocols to be used by Neon.

Patch by: djh
(Tweaked by me.)
Reviewed by: me
             rooneg


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@861605 13f79535-47bb-0310-9956-ffa450edef68