Fix a hang during character set translation by ensuring that we can always make progress through a string.
If we're converting to a multibyte encoding (such as UTF-8), we need
to ensure that we always have at least the space required to convert
any valid character in the source encoding, otherwise we won't make
any progress and will hang.

For conversions to UTF-8, that means we always need to have at least
three spare bytes in the output buffer (for characters in the BMP),
and that sounds like a good assumption for other encodings as well.

Found by: Vlad Skvortsov <vss@73rus.com>

* subversion/libsvn_subr/utf.c
  (convert_to_stringbuf): Expand the destination buffer using a 1:3
    ratio whenever we have less than 3 bytes spare, ensuring we can
    always make progress through a string.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@863566 13f79535-47bb-0310-9956-ffa450edef68