When determining a copy source with svn_client__get_copy_source, notice copied parents of our target path.
For example, if you have "/branches/foo/bar", and "/branches/baz/bar",
and "baz" was a copy of "foo", we should detect the copy when running
'log' on "bar".  Consider a revision like:

  $ svn log -v -r2 branches/baz/bar
  ------------------------------------------------------------------------
  r2 | dlr | 2007-05-25 19:11:55 -0700 (Fri, 25 May 2007) | 1 line
  Changed paths:
     A /branches/baz (from /branches/foo:1)
  
  create branch
  ------------------------------------------------------------------------

The changed paths hash passed to a log receiver will contain
/branches/baz, not /branches/baz/bar, so in order to find the copy
source, we need to check if the changed path is a parent of the give
path.

This code (incorrectly) assumes that we'll never see both "baz" and
"baz/bar" in the log output.  To handle the most general case, we will
need to use logic similar to svn_fs_closest_copy (or just expose
svn_fs_closest_copy over the RA layer).

* subversion/libsvn_client/log.c
  (copyfrom_info_receiver): Detect copied parents of BATON->target_path,
   and record copy source appropriately.

Reviewed by: kameshj


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@865254 13f79535-47bb-0310-9956-ffa450edef68