*** Reverted in r26142.  Will be handled under the auspices of issue #571. ***
Change the expected output for the scenario of removing a directory,
commiting the removal, coping another directory in place of the old
directory, commiting that, then performing a reverse merge to restore
the initial directory from 'A ' (add of the directory) to 'R '
(replace of the directory).

Eric H. suggests fixing this by changing the merge callbacks maintain
some state in their baton to track whether a remove has already
occurred on a path which is getting a subsequent add.

* subversion/tests/cmdline/revert_tests.py
  (status_of_missing_dir_after_revert_replaced_with_history_dir): Expect
   'R ' instead of 'A ' for the reverse merge which replaces 'A/D/G'.
  (test_list): Mark
   status_of_missing_dir_after_revert_replaced_with_history_dir as XFAIL.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@866210 13f79535-47bb-0310-9956-ffa450edef68